refactor cypress setToken into helper fxn
